Prepare sequential solution for parallelism II
Change the way of how found solutions are passed and stored.
Do not pass found solutions as return values, store them in the instance
variable intead.
This is because solutions will be found in parallel and thus it is
better for tasks to be able to store them by just updating the variable.